% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbss.R
\name{nbss}
\alias{nbss}
\title{Normalised Biomass Size Spectrum}
\usage{
nbss(x, type = c("biomass", "abundance"), base = 10, binwidth = 0.1)
}
\arguments{
\item{x}{vector of biomasses, biovolumes, or lengths.}

\item{type}{whether to compute a biomass/biovolume or abundance spectrum.}

\item{base}{base of the logarithm for the computation of bins.}

\item{binwidth}{width of bins in log10 scale.}
}
\value{
a data.frame with columns
}
\description{
Normalised Biomass Size Spectrum
}
\examples{
# Biovolume spectrum
ss <- nbss(uvp$volume_mm3)
head(ss)
autoplot(ss) + labs(
  x=expression("Biovolume (mm"^3*")"),
  y="Normalised biovolume"
)
}
