% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbss.R
\name{nbss}
\alias{nbss}
\title{Normalised Biomass Size Spectrum}
\usage{
nbss(x, type = c("biomass", "abundance"), base = 10, binwidth = 0.1)
}
\arguments{
\item{x}{vector of biomasses, biovolumes, or lengths.}

\item{type}{whether to compute a biomass/biovolume or abundance spectrum.}

\item{base}{base of the logarithm for the computation of bins.}

\item{binwidth}{width of bins in log10 scale.}
}
\value{
a data.frame with columns
\itemize{
\item \code{bin_log} value of the bin center in log scale;
\item \code{bin} value of the bin center in original scale;
\item \code{binwidth} width of the bin center in original scale;
\item \code{y} sum of biomass/biovolume for a biomass spectrum; count of the objects in the bin for an abundance spectrum;
\item \code{norm_y} \code{y/binwidth}
}
}
\description{
Normalised Biomass Size Spectrum
}
\examples{
# Biovolume spectrum
ss <- nbss(uvp$volume_mm3)
head(ss)
autoplot(ss) + labs(
  x=expression("Biovolume (mm"^3*")"),
  y="Normalised biovolume"
)
# Abundance spectrum
ss <- nbss(uvp$length_mm, binwidth=0.05)
autoplot(ss) + labs(x="Length (mm)")
}
